+++ src/Fable.Transforms/Fable2Babel.fs
@@ -334,6 +334,7 @@ module Reflection =
         | Fable.LambdaType(argType, returnType) -> genericTypeInfo "lambda" [ argType; returnType ]
         | Fable.DelegateType(argTypes, returnType) -> genericTypeInfo "delegate" [ yield! argTypes; yield returnType ]
         | Fable.Tuple(genArgs, _) -> genericTypeInfo "tuple" genArgs
+        | Fable.Nullable(genArg, isStruct) -> transformTypeInfoFor purpose com ctx r genMap genArg
         | Fable.Option(genArg, _) -> genericTypeInfo "option" [ genArg ]
         | Fable.Array(genArg, _) -> genericTypeInfo "array" [ genArg ]
         | Fable.List genArg -> genericTypeInfo "list" [ genArg ]
@@ -345,7 +346,6 @@ module Reflection =
             List.zip (fieldNames |> Array.toList) genArgs
             |> List.map (fun (k, t) -> Expression.arrayExpression [| Expression.stringLiteral (k); t |])
             |> libReflectionCall com ctx None "anonRecord"
-        | Fable.Nullable typ -> transformTypeInfoFor purpose com ctx r genMap typ
         | Fable.DeclaredType(entRef, genArgs) ->
             let fullName = entRef.FullName
 
@@ -496,7 +496,7 @@ module Reflection =
         | Fable.MetaType -> jsInstanceof (libValue com ctx "Reflection" "TypeInfo") expr
         | Fable.Option _ -> warnAndEvalToFalse "options" // TODO
         | Fable.GenericParam _ -> warnAndEvalToFalse "generic parameters"
-        | Fable.Nullable typ -> transformTypeTest com ctx range expr typ
+        | Fable.Nullable(genArg, _isStruct) -> transformTypeTest com ctx range expr genArg
         | Fable.DeclaredType(ent, genArgs) ->
             match ent.FullName with
             | Types.idisposable ->
@@ -611,6 +611,7 @@ module Annotation =
         | Fable.Regex -> makeAliasTypeAnnotation com ctx "RegExp"
         | Fable.Number(BigInt, _) -> makeAliasTypeAnnotation com ctx "bigint"
         | Fable.Number(kind, _) -> makeNumericTypeAnnotation com ctx kind
+        | Fable.Nullable(genArg, _) -> makeNullableTypeAnnotation com ctx genArg
         | Fable.Option(genArg, _) -> makeOptionTypeAnnotation com ctx genArg
         | Fable.Tuple(genArgs, _) -> makeTupleTypeAnnotation com ctx genArgs
         | Fable.Array(genArg, kind) -> makeArrayTypeAnnotation com ctx genArg kind
@@ -620,7 +621,6 @@ module Annotation =
         | Fable.DelegateType(argTypes, returnType) -> makeFunctionTypeAnnotation com ctx typ argTypes returnType
         | Fable.AnonymousRecordType(fieldNames, fieldTypes, _isStruct) ->
             makeAnonymousRecordTypeAnnotation com ctx fieldNames fieldTypes
-        | Fable.Nullable typ -> makeNullableTypeAnnotation com ctx typ
         | Replacements.Util.Builtin kind -> makeBuiltinTypeAnnotation com ctx typ kind
         | Fable.DeclaredType(entRef, genArgs) -> com.GetEntity(entRef) |> makeEntityTypeAnnotation com ctx genArgs
 
@@ -884,8 +884,6 @@ module Annotation =
 
     let makeEntityTypeAnnotation com ctx genArgs (ent: Fable.Entity) =
         match genArgs, ent with
-        | [ genArg ], EntFullName Types.nullable -> makeNullableTypeAnnotation com ctx genArg
-
         | _, Patterns.Try (tryNativeOrFableLibraryInterface com ctx genArgs) ta -> ta
 
         | _, Patterns.Try (Lib.tryJsConstructorFor Annotation com ctx) entRef ->
@@ -1981,6 +1979,16 @@ module Util =
                 | Fable.Value(Fable.Null _, _), e
                 | e, Fable.Value(Fable.Null _, _) ->
                     com.TransformAsExpr(ctx, e) |> makeNullCheck range (op = BinaryEqual)
+                | ExprType(Fable.Nullable _), ExprType(Fable.Nullable _) ->
+                    Replacements.Util.Helper.LibCall(
+                        com,
+                        "Util",
+                        "nullableEquals",
+                        Fable.Boolean,
+                        [ e1; e2 ],
+                        ?loc = range
+                    )
+                    |> transformAsExpr com ctx
                 | ExprType(Fable.MetaType), _ ->
                     let e =
                         Replacements.Util.Helper.LibCall(