+++ src/fable-library-ts/Option.ts
@@ -52,7 +52,7 @@ export class Some<T> {
   }
 }
 
-export function nullableValue<T>(x: Nullable<T>): T {
+export function nonNullValue<T>(x: Nullable<T>): T {
   if (x == null) {
     throw new Error("Nullable has no value");
   } else {