+++ src/Fable.AST/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+* [All] Support for Nullable Reference Types (by @ncave and @MangelMaxime)
+
 ## 5.0.0-beta.1 - 2025-02-16
 
 ### Added