+++ src/Fable.Cli/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * [JS/TS] Fix #3533: Add directives prologues supports (by @MangelMaxime)
-* [TS] Initial support for Nullable Reference Types (by @ncave)
+* [JS/TS] Support for Nullable Reference Types (by @ncave and @MangelMaxime)
 
 ### Changed